// src/interfaces/api.interface.ts
export interface baseRes<D = {}> {
	success: boolean
	code: number
	message: string
	data: D //返回的数据字段
}

export interface PageData<D = {}> {
	tempGroup: any
	bindList: any
	filePreviewUrl: string
	fileSaveUrl: string
	currentPage: number
	pageNum: number
	pageSize: number
	totalPage: number
	count: number
	list: D[]
	fileHost: string
	helpVideo: {
		videoUrl: string
		coverUrl: string
	}
}
// ...
export interface styleList {
	styleId?: string
}
// ...
export interface formListItem {
	id: number
	applyTime: number
	area: number
	handTime: string
	handler: string
	handlerId: string
	status: string
	telephone: string
	type: string
	remarks?: string
	tempId?: string
}

export interface formTypeItem {
	title?: string
	id?: number
	key: number | string
	value: string
	tempId?: string
}
export interface sortTypeItem {
	createLink: string
	description: string
	hoverIcon: string
	icon: string
	id: string
	moreLink: string
	name: string
	num: number
	sort: number
	tag: string
}

export interface bannerListItem {
	commentNum: number
	contentUrl: string
	createTs: number
	enable: boolean
	endTime: number
	id: number
	imgUrl: string
	sort: number
	startTime: number
	title: string
	type: string
	username: string
	description?: string
	industryId?: number
}

export interface bindcaseItem {
	thumb: string
}

export interface orderInfo {
	createTs: number
	fee: string
	outTradeNo: string
	sceneTemplateId: string
	sceneTemplateName: string
	status: string
	transactionType: string
	validSz: number
}

export interface caseListItem {
	caseSort: number
	caseStatus: number
	createTs: number
	id: number
	sceneId: string
	styleId: string
	styleName: string
	sz: number
	tempId: string
	tempName: string
	thumb: string
	typeId: number
	typeName: string
	validSz: number
	recomList: any
	styleList: any
  recomName: string
  recomType: number
}

export interface authInfoData {
	authorities: Authority[]
	userRoles: UserRole[]
}

export interface Authority {
	id: number
	parentId: number
	action: string
	icon: string
	name: string
	url?: string
	sort: number
}

export interface UserRole {
	roleId: number
	roleName: string
}

export interface hcaseListItem {
	id: string
	name: string
	url: string
	indexTradeId: number
	sort: number
	image?: string
}

export interface otherPermissionsItem {
	createTs: number
	description: string
	id: number
	permissionName: string
}

export interface sortListItem {
	createLink: string
	description: string
	hoverIcon: string
	icon: string
	id: string
	moreLink: string
	name: string
	num: number
	sort: number
	tag: string
}
// ...
export interface upgradeListItem {
	id: string
	createTs: string
	createdBy: string
	updateTs: string
	updatedBy: string
	title: string
	describe: string
	status: number
	deleted: number
	sort: number
}

export interface upgradeEditItem {
	info: upgradeInfoItem
	detailsList: upgradeDetailsListItem[]
}

export interface upgradeInfoItem {
	createTs: number
	createdBy: string
	deleted: number
	describe: string
	id: string
	sort: number
	status: number
	title: string
	updateTs: number
	updatedBy: string
}

export interface upgradeDetailsListItem {
	createTs: number
	createdBy: string
	describe: string
	fileUrl: string | any
	id: string
	linkUrl: string
	name: string
	sort: number
	updateTs: number
	updatedBy: string
	upgradeId: string
}
// ...
export interface polymerizationInfo {
	imgUrl: string
}

export interface loginListItem {
	id: string
	name: string
	thumb: string
	styleFile: string
	relatedTopic: boolean
	rackStatus: boolean
	sort: number
}

export interface visitorsListItem {
	tempId: string
	tempName: string
	owner: string
	company: string
	realName: string
	telephone: string
	remark: string
	createTs: string
}