// src/interfaces/params.interface.ts
export interface PageParams {
	pageNum?: number
	pageSize?: number
	keyword?: string
}
// ...
export interface hallListParams {
	sceneId?: string
	sceneName?: string
	srcType?: string
}
// ...
export interface currencyParams {
	keywords: string
	pageNum: number
	pageSize: number
}
// ...
export interface formsParams {
	type?: string
	status?: string
	keyword?: string
}

export interface handleFormParams {
	id: number
	version?: number
	status: number
	remark: string
}
export interface bindcaseParams {
	tempId: string
}
export interface bannerListParams {
	type?: string
}

export interface bannerParams {
	title: string
	imgUrl: string
	content?: string
	contentUrl: string
	type: number
	sort: number
	id: number
	industryId?: number
	remark?: string
}

export interface orderListParams {
	belonger: string
	transactionType: number
	keyword: string
}

export interface rechargeListParams {
	keyword: string
}

export interface caseParams {
	id?: string
	styleId?: string
	typeId?: string
	tempId?: string
	tempName?: string
	caseStatus?: string
	tempRecomList?: string
	styleIdList?: string
}

export interface caseListParams {
	keyword?: string
}
// ...
export interface polymerizationParams {
	keyword: String
}
export interface addSale {
	active?: boolean
	id?: number
	name: string
	telephone: string
}
export interface couponList {
	salesmanId?: string
	type?: number
	couponName?: string
	status: number
}
export interface addPolymerization {
	title: string
	remark: string
	imgUrl: string
	contentUrl: string
	sort: number
}
export interface addDiscount {
	salesmanId: string
	name: string
	discountAmount: number
	highestDeductionAmount: number
	type: number
	validStartTime: string
	validEndTime: string
	amount: number
	desc: string
	id?: string
}

export interface logListParams {
	keyword: string
	creator: string
	startTime: string
	endTime: string
	businessType: string
}

export interface upgradeAddInfoParams {
	info: upgradeInfoParams
	detailsList: upgradeDetailsListParams[]
}

export interface upgradeInfoParams {
	title: string
	status: number
}

export interface upgradeDetailsListParams {
	name: string
	describe: string
	linkUrl: string
	fileUrl: string
}
// ...
export interface visitorsListParams {
	tempIdOrName: string
	telephoneOrName: string
	startTime: string
	endTime: string
}

export interface exportParams {
	params: number[]
	keyword: string
	startTime: string
	endTime: string
}

export interface exportVisitorsParams {
	tempIdOrName: string
	telephoneOrName: string
	startTime: string
	endTime: string
}